RN-976: Refactored useQrCode to be a bit more DRY
