+++ packages/ui-components/src/components/QrCode/index.ts
@@ -4,5 +4,5 @@
  */
 
 export * from './QrCodeVisual';
-export * from './useQrCodeCanvas';
+export * from './useQrCode';
 export * from './QrCodeImage';
